Add files via upload
This is the code for ODE system of particles just with alignment forces, without the strong noise of the paper. The code is working, in particular when we confront the result with the old code we have the same result. In this version of the code we compute and plot the average trajectory of particles